% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-features.r
\name{BinFeaturesPlot_MultipleData}
\alias{BinFeaturesPlot_MultipleData}
\title{Visualization of characteristic binary features of multiple data sets}
\usage{
BinFeaturesPlot_MultipleData(
  leadCpds,
  orderLab,
  features = list(),
  data = list(),
  validate = NULL,
  colorLab = NULL,
  nrclusters = NULL,
  cols = NULL,
  name = c("Data1", "Data2"),
  colors1 = c("gray90", "blue"),
  colors2 = c("gray90", "green"),
  margins = c(5.5, 3.5, 0.5, 5.5),
  cexB = 0.8,
  cexL = 0.8,
  cexR = 0.8,
  spaceNames = 0.2,
  plottype = "new",
  location = NULL
)
}
\arguments{
\item{leadCpds}{A character vector with the names of the objects in a first group, i.e., the group for which the specified features are characteristic. Default is NULL.}

\item{orderLab}{A character vector with the order of the objects. Default is NULL.}

\item{features}{A list with as elements character vectors with the names of the features to be visualized for each data set. Default is NULL.}

\item{data}{A list with the different data sets. Default is NULL.}

\item{validate}{Optional. A list with validation data sets. If a feature has a validation reference, these are added in a red colour. Default is NULL.}

\item{colorLab}{Optional. A clustering object if the objects are to be coloured accoring to their clustering order. Default is NULL.}

\item{nrclusters}{Optional. The number of clusters to divide the dendrogram of ColorLab. Default is NULL.}

\item{cols}{Optional. A character vector with the colours of the different clusters. Default is NULL.}

\item{name}{A character string with the names of the data sets. Default is c("Data1" ,"Data2") for two data sets.}

\item{colors1}{A character vector with the colours to indicate the presence (first element) or the absence of the features for the objects in LeadCpds. Default is c('gray90','blue').}

\item{colors2}{A character vector with the colours to indicate the presence (first element) or the absence of the features for the objects in the remaining objects. Default is c('gray90','green').}

\item{margins}{A vector with the margings of the plot. Default is c(5.5,3.5,0.5,5.5).}

\item{cexB}{The font size of the labels on the bottom: the object labels. Default is 0.80.}

\item{cexL}{The font size of the labels on the left: the data labels. Default is 0.80.}

\item{cexR}{The font size of the labels on the right: the feature labels. Default is 0.80.}

\item{spaceNames}{A percentage of the height of the figure to be reserved for the names of the objects. Default is 0.20.}

\item{plottype}{Should be one of "pdf","new" or "sweave". If "pdf", a
location should be provided in "location" and the figure is saved there. If
"new" a new graphic device is opened and if "sweave", the figure is made
compatible to appear in a sweave or knitr document, i.e. no new device is
opened and the plot appears in the current device or document. Default is "new".}

\item{location}{Optional. If plottype is "pdf", a location should be provided in
"location" and the figure is saved there. Default is NULL.}
}
\value{
A plot visualizing the presence and absence of the characteristic binary features across multiple data sets.
}
\description{
A tool to visualize characteristic binary features of a set of objects in comparison with the remaining objects for multiple data sets. The result is a matrix with coloured cells. Columns represent
objects and rows represent the specified features. A feature which is present is give a coloured cell while an absent feature is represented by a grey cell. The labels on the right indicate the names of the features while the labels on the bottom are the names of the objects.
}
\examples{
\dontrun{
data(fingerprintMat)
data(targetMat)

MCF7_F = Cluster(fingerprintMat,type="data",distmeasure="tanimoto",normalize=FALSE,
method=NULL,clust="agnes",linkage="flexible",gap=FALSE,maxK=55,StopRange=FALSE)

Comps=FindCluster(list(MCF7_F),nrclusters=10,select=c(1,8))

MCF7_Char=CharacteristicFeatures(List=NULL,Selection=Comps,binData=
list(fingerprintMat,targetMat),datanames=c("FP","TP"),nrclusters=NULL,
topC=NULL,sign=0.05,fusionsLog=TRUE,weightclust=TRUE,names=c("FP","TP"))

FeatFP=MCF7_Char$Selection$Characteristics$FP$TopFeat$Names[c(1:10)]
FeatTP=MCF7_Char$Selection$Characteristics$TP$TopFeat$Names[c(1:10)]

BinFeaturesPlot_MultipleData(leadCpds=Comps,orderLab=MCF7_Char$Selection$
objects$OrderedCpds,features=list(FeatFP,FeatTP),data=list(fingerprintMat,targetMat),
validate=NULL,colorLab=NULL,nrclusters=NULL,cols=NULL,name=c("FP","TP"),colors1=
c('gray90','blue'),colors2=c('gray90','green'),margins=c(5.5,3.5,0.5,5.5),cexB=0.80,
cexL=0.80,cexR=0.80,spaceNames=0.20,plottype="new",location=NULL)
}
}
