% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-characterize.r
\name{CharacteristicFeatures}
\alias{CharacteristicFeatures}
\title{Determine the characteristic features of clusters}
\usage{
CharacteristicFeatures(
  List,
  Selection = NULL,
  binData = NULL,
  contData = NULL,
  datanames = NULL,
  nrclusters = NULL,
  sign = 0.05,
  topChar = NULL,
  fusionsLog = TRUE,
  weightclust = TRUE,
  names = NULL
)
}
\arguments{
\item{List}{A list of clustering outputs to be compared. The first element of
the list is used as the reference in \code{ReorderToReference}.}

\item{Selection}{If a selection of objects is provided, the function only
investigates the features of this selection. Default is NULL.}

\item{binData}{A list of the binary feature data matrices. Default is NULL.}

\item{contData}{A list of continuous feature data matrices. Default is NULL.}

\item{datanames}{A vector with the names of the data matrices. Default is NULL.}

\item{nrclusters}{The number of clusters to cut the dendrogram in. Default is NULL.}

\item{sign}{The significance level. Default is 0.05.}

\item{topChar}{The number of top characteristics to return. If NULL, only the
significant characteristics are saved. Default is NULL.}

\item{fusionsLog}{Logical. Indicator for the fusion of clusters. Default is TRUE.}

\item{weightclust}{Logical. For the outputs of CEC, WeightedClust or
WeightedSimClust, if TRUE only the result of the Clust element is considered.
Default is TRUE.}

\item{names}{Optional. Names of the methods. Default is NULL.}
}
\value{
A list with for each method the found (top) characteristics of the
feature data per cluster.
}
\description{
\code{CharacteristicFeatures} identifies the features that are characteristic
for the objects of each cluster of one or more clustering results. Binary
features are evaluated with Fisher's exact test, continuous features with the
t-test. Multiplicity correction is included.
}
\examples{
\dontrun{
data(fingerprintMat)
data(targetMat)

MCF7_F = Cluster(fingerprintMat,type="data",distmeasure="tanimoto",normalize=FALSE,
method=NULL,clust="agnes",linkage="flexible",gap=FALSE,maxK=55,StopRange=FALSE)
MCF7_T = Cluster(targetMat,type="data",distmeasure="tanimoto",normalize=FALSE,
method=NULL,clust="agnes",linkage="flexible",gap=FALSE,maxK=55,StopRange=FALSE)

L=list(MCF7_F,MCF7_T)
MCF7_Char=CharacteristicFeatures(List=L,Selection=NULL,binData=list(fingerprintMat),
contData=NULL,datanames=c("FP"),nrclusters=7,topChar=20)
}
}
\references{
Perualila-Tan N. et al. (2016). Weighted similarity-based clustering
of chemical structures and bioactivity data in early drug discovery.
}
