% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-characterize.r
\name{ChooseCluster}
\alias{ChooseCluster}
\title{Interactive plot to determine DE Genes and DE features for a specific cluster}
\usage{
ChooseCluster(
  Interactive = TRUE,
  leadCpds = NULL,
  clusterResult = NULL,
  colorLab = NULL,
  binData = NULL,
  contData = NULL,
  datanames = c("FP"),
  geneExpr = NULL,
  topChar = 20,
  topG = 20,
  sign = 0.05,
  nrclusters = NULL,
  cols = NULL,
  n = 1
)
}
\arguments{
\item{Interactive}{Logical. Whether an interactive plot should be made. Defaults to TRUE.}

\item{leadCpds}{A list of the objects of the clusters of interest. Defaults to NULL.}

\item{clusterResult}{The output of one of the aggregated cluster functions. Default is NULL.}

\item{colorLab}{The clustering result the dendrogram should be colored after. Default is NULL.}

\item{binData}{A list of the binary feature data matrices. Default is NULL.}

\item{contData}{A list of continuous data sets of the objects. Default is NULL.}

\item{datanames}{A vector with the names of the data matrices. Default is c("FP").}

\item{geneExpr}{A gene expression matrix, may also be an ExpressionSet. Default is NULL.}

\item{topChar}{The number of top characteristics to return. Default is 20.}

\item{topG}{The number of top genes to return. Default is 20.}

\item{sign}{The significance level. Default is 0.05.}

\item{nrclusters}{Optional. The number of clusters to cut the dendrogram in. Default is NULL.}

\item{cols}{The colors to use in the dendrogram. Default is NULL.}

\item{n}{The number of clusters one wants to identify by a mouse click. Default is 1.}
}
\value{
A list with one element per cluster of interest with elements
objects, Characteristics and (optionally) Genes.
}
\description{
If desired, the function produces a dendrogram of a clustering result. One
or multiple clusters can be indicated by a mouse click. From these clusters
DE genes and characteristic features are determined. It is also possible to
provide the objects of interest without producing the plot.
}
\details{
This function may require the suggested packages 'a4Base', 'a4Core'
and 'limma' when a gene expression matrix is supplied.
}
\examples{
\dontrun{
data(fingerprintMat)
data(targetMat)
data(geneMat)

MCF7_F = Cluster(fingerprintMat,type="data",distmeasure="tanimoto",normalize=FALSE,
method=NULL,clust="agnes",linkage="flexible",gap=FALSE,maxK=55,StopRange=FALSE)
MCF7_T = Cluster(targetMat,type="data",distmeasure="tanimoto",normalize=FALSE,
method=NULL,clust="agnes",linkage="flexible",gap=FALSE,maxK=55,StopRange=FALSE)

MCF7_Interactive=ChooseCluster(Interactive=TRUE,leadCpds=NULL,clusterResult=MCF7_T,
colorLab=MCF7_F,binData=list(fingerprintMat),datanames=c("FP"),geneExpr=geneMat,
topChar = 20, topG = 20,nrclusters=7,n=1)
}
}
