% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-dendro.r
\name{ColorPalette}
\alias{ColorPalette}
\title{Create a color palette to be used in the plots}
\usage{
ColorPalette(colors = c("red", "green"), ncols = 5)
}
\arguments{
\item{colors}{A vector containing the colors of choice}

\item{ncols}{The number of colors to be specified. If higher than the number
of colors, it specifies colors in the region between the given colors.}
}
\value{
A vector containing the hex codes of the chosen colors.
}
\description{
In order to facilitate the visualization of the influence of the different
methods on the clustering of the objects, colours can be used. The function
\code{ColorPalette} is able to pick out as many colours as there are
clusters. This is done with the help of the \code{ColorRampPalette} function
of the grDevices package
}
\examples{

Colors1<-ColorPalette(c("cadetblue2","chocolate","firebrick2",
"darkgoldenrod2", "darkgreen","blue2","darkorchid3","deeppink2"), ncols=8)
}
