% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compare.r
\name{CompareInteractive}
\alias{CompareInteractive}
\title{Interactive comparison of single and multiple source clustering results}
\usage{
CompareInteractive(
  ListM,
  ListS,
  nrclusters = NULL,
  cols = NULL,
  fusionsLogM = FALSE,
  fusionsLogS = FALSE,
  weightclustM = FALSE,
  weightclustS = FALSE,
  namesM = NULL,
  namesS = NULL,
  marginsM = c(2, 2.5, 2, 2.5),
  marginsS = c(8, 2.5, 2, 2.5),
  Interactive = TRUE,
  n = 1
)
}
\arguments{
\item{ListM}{A list of the outputs from the multiple source clusterings to be compared.}

\item{ListS}{A list of the outputs from the single source clusterings to be compared.}

\item{nrclusters}{The number of clusters to cut the dendrogram in. Default is NULL.}

\item{cols}{A character vector with the names of the colours. Default is NULL.}

\item{fusionsLogM}{The fusionsLog parameter for the elements in ListM. Default is FALSE.}

\item{fusionsLogS}{The fusionsLog parameter for the elements in ListS. Default is FALSE.}

\item{weightclustM}{The weightclust parameter for the elements in ListM. Default is FALSE.}

\item{weightclustS}{The weightclust parameter for the elements in ListS. Default is FALSE.}

\item{namesM}{Optional. Names of the multiple source clusterings. Default is NULL.}

\item{namesS}{Optional. Names of the single source clusterings. Default is NULL.}

\item{marginsM}{Optional. Margins to be used for the ListM plot. Default is c(2,2.5,2,2.5).}

\item{marginsS}{Optional. Margins to be used for the ListS plot. Default is c(8,2.5,2,2.5).}

\item{Interactive}{Optional. Do you want an interactive plot? Default is TRUE.}

\item{n}{The number of methods/clusters you want to identify. Default is 1.}
}
\value{
A plot of the comparison of the elements of ListM, on which multiple clusters
and/or methods can be identified to compare them to the elements in ListS.
}
\description{
The function \code{CompareInteractive} produces an interactive plot of the
multiple source clustering results in \code{ListM}. By identifying a cluster
or a method, a comparison with the single source clustering results in
\code{ListS} is shown in a new plotting window.
}
\examples{
\dontrun{
data(fingerprintMat)
data(targetMat)
data(Colors1)

MCF7_F = Cluster(fingerprintMat,type="data",distmeasure="tanimoto",normalize=FALSE,
method=NULL,clust="agnes",linkage="flexible",gap=FALSE,maxK=55,StopRange=FALSE)
MCF7_T = Cluster(targetMat,type="data",distmeasure="tanimoto",normalize=FALSE,
method=NULL,clust="agnes",linkage="flexible",gap=FALSE,maxK=55,StopRange=FALSE)

L=list(fingerprintMat,targetMat)

MCF7_W=WeightedClust(List=L,type="data",distmeasure=c("tanimoto","tanimoto"),
normalize=c(FALSE,FALSE),method=c(NULL,NULL),weight=seq(1,0,-0.1),weightclust=0.5,
clust="agnes",linkage="ward",StopRange=FALSE)

ListM=list(MCF7_W)
namesM=c(seq(1.0,0.0,-0.1))

ListS=list(MCF7_F,MCF7_T)
namesS=c("FP","TP")

CompareInteractive(ListM,ListS,nrclusters=7,cols=Colors1,fusionsLogM=FALSE,
fusionsLogS=FALSE,weightclustM=FALSE,weightclustS=TRUE,namesM,namesS,
marginsM=c(2,2.5,2,2.5),marginsS=c(8,2.5,2,2.5),Interactive=TRUE,n=1)
}
}
