% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-dendro.r
\name{ComparePlot}
\alias{ComparePlot}
\title{Comparison of clustering results over multiple methods}
\usage{
ComparePlot(
  List,
  nrclusters = NULL,
  cols = NULL,
  fusionsLog = FALSE,
  weightclust = FALSE,
  names = NULL,
  margins = c(8.1, 3.1, 3.1, 4.1),
  circle = FALSE,
  canvaslims = c(-1, 1, -1, 1),
  Highlight = NULL,
  substr = NULL,
  cex.highlight = 1,
  cex.labels = 0.7,
  trackheightdend = 0.5,
  plottype = "new",
  location = NULL
)
}
\arguments{
\item{List}{A list of the outputs from the methods to be compared. The first
element of the list is used as the reference in \code{ReorderToReference}.}

\item{nrclusters}{The number of clusters to cut the dendrogram in. Default is NULL.}

\item{cols}{A character vector with the colours to be used. Default is NULL.}

\item{fusionsLog}{Logical. To be handed to \code{ReorderToReference}. Default is FALSE.}

\item{weightclust}{Logical. To be handed to \code{ReorderToReference}. Default is FALSE.}

\item{names}{Optional. Names of the methods to be used as labels. Default is NULL.}

\item{margins}{Optional. Margins for the plot. Default is c(8.1,3.1,3.1,4.1).}

\item{circle}{Logical. Whether the figure should be circular (TRUE) or a rectangle (FALSE). Default is FALSE.}

\item{canvaslims}{The limits for the circular dendrogram. Default is c(-1.0,1.0,-1.0,1.0).}

\item{Highlight}{Optional. A list of character vectors of objects to be highlighted. Default is NULL.}

\item{substr}{Optional. A vector of length two giving start and stop positions to shorten labels. Default is NULL.}

\item{cex.highlight}{Magnification for the highlight text. Default is 1.}

\item{cex.labels}{Magnification for the labels. Default is 0.7.}

\item{trackheightdend}{The height of the dendrogram track in the circular plot. Default is 0.5.}

\item{plottype}{Should be one of "pdf","new" or "sweave". Default is "new".}

\item{location}{Optional. If plottype is "pdf", a location should be provided. Default is NULL.}
}
\value{
A plot which translates the matrix output of \code{ReorderToReference}.
}
\description{
A visual comparison of the clustering results of several methods.
The function relies on \code{ReorderToReference} and \code{ColorsNames} and
renders the resulting matrix either as a rectangular heatmap (using the
\code{plotrix} package) or in a circular format (using the \code{circlize}
package).
}
