% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-dendro.r
\name{Cyclogram}
\alias{Cyclogram}
\title{Comparison of clustering results over multiple results in circular format}
\usage{
Cyclogram(
  List,
  nrclusters = NULL,
  cols = NULL,
  fusionsLog = FALSE,
  weightclust = FALSE,
  names = NULL,
  canvaslims = c(-1, 1, -1, 1),
  margins = c(8.1, 3.1, 3.1, 4.1),
  Highlight = NULL,
  cex.highlight = 1,
  substr = NULL,
  cex.labels = 0.7,
  plottype = "new",
  location = NULL
)
}
\arguments{
\item{List}{A list of the outputs from the methods to be compared. The first
element of the list will be used as the reference in
\code{ReorderToReference}.}

\item{nrclusters}{The number of clusters to cut the dendrogram in. Default is NULL.}

\item{cols}{A character vector with the colours to be used. Default is NULL.}

\item{fusionsLog}{Logical. To be handed to \code{ReorderToReference}: indicator for the fusion of clusters. Default is TRUE}

\item{weightclust}{Logical. To be handed to \code{ReorderToReference}: to be used for the outputs of CEC,
WeightedClust or WeightedSimClust. If TRUE, only the result of the Clust element is considered. Default is TRUE.}

\item{names}{Optional. Names of the methods to be used as labels for the
columns. Default is NULL.}

\item{canvaslims}{The limits for the circular dendrogam. Default is c(-1.0,1.0,-1.0,1.0).}

\item{margins}{Optional. Margins to be used for the plot. Default is c(8.1,3.1,3.1,4.1).}

\item{Highlight}{Optional. A list of character vectors of objects to be highlighted. Default is NULL.}

\item{cex.highlight}{Magnification for the highlight text. Default is 1.}

\item{substr}{Optional. A vector of length two giving start and stop positions to shorten labels. Default is NULL.}

\item{cex.labels}{Magnification for the labels. Default is 0.7.}

\item{plottype}{Should be one of "pdf","new" or "sweave". If "pdf", a
location should be provided in "location" and the figure is saved there. If
"new" a new graphic device is opened and if "sweave", the figure is made
compatible to appear in a sweave or knitr document, i.e. no new device is
opened and the plot appears in the current device or document. Default is "new".}

\item{location}{Optional. If plottype is "pdf", a location should be provided in
"location" and the figure is saved there. Default is NULL.}
}
\value{
A plot which translates the matrix output of the function
\code{ReorderToReference} into a circular format.
}
\description{
A visual comparison of all methods is handy to see which objects will
always cluster together independent of the applied methods. To this aid the
function \code{Cyclogram} has been written. The function relies on methods
of the \code{circlize} package.
}
\details{
This function makes use of the functions \code{ReorderToReference} and
\code{Colorsnames}. Given a list with the outputs of several methods, the
first step is to call upon \code{ReorderToReference} and to produce a
matrix of which the columns are ordered according to the ordering of the
objects of the first method in the list. Each cell represent the number of
the cluster the object belongs to for a specific method indicated by the
rows. The clusters are arranged in such a way that these correspond to that
one cluster of the referenced method that they have the most in common with.
The \code{circlize} package is used to visualize the matrix is a circular format.
The inner element of the circle is the first element of the List parameter, the second inner
element is the second element of the list and so on. The object names are written on the
circular dendrogram portayed in the center of the circle.
}
