% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{DiffGenes}
\alias{DiffGenes}
\title{Find differentially expressed genes}
\usage{
DiffGenes(
  List,
  Selection = NULL,
  geneExpr = NULL,
  nrclusters = NULL,
  method = "limma",
  sign = 0.05,
  topG = NULL,
  fusionsLog = TRUE,
  weightclust = TRUE,
  names = NULL
)
}
\arguments{
\item{List}{A list of the clustering outputs to be compared. The first
element of the list will be used as the reference in
\code{ReorderToReference}.}

\item{Selection}{If differential gene expression should be investigated for
a specific selection of objects, this selection can be provided here.
Selection can be of the type "character" (names of the objects) or
"numeric" (the number of specific cluster). Default is NULL.}

\item{geneExpr}{The gene expression matrix or ExpressionSet of the objects.
The rows should correspond with the genes.}

\item{nrclusters}{Optional. The number of clusters to cut the dendrogram in.
Default is NULL.}

\item{method}{The method to applied to look for DE genes. For now, only the
limma method is available. Default is "limma".}

\item{sign}{The significance level to be handled. Default is 0.05.}

\item{topG}{Overrules sign. The number of top genes to be shown. Default is NULL.}

\item{fusionsLog}{Logical. To be handed to \code{ReorderToReference}: indicator for the fusion of clusters. Default is TRUE}

\item{weightclust}{Logical. To be handed to \code{ReorderToReference}: to be used for the outputs of CEC,
WeightedClust or WeightedSimClust. If TRUE, only the result of the Clust element is considered. Default is TRUE.}

\item{names}{Optional. Names of the methods. Default is NULL.}
}
\value{
The result is a list with an element per method. Each element is
again a list with for each cluster the found differentially expressed genes.
}
\description{
The \code{DiffGenes} function looks for differentially expressed
genes for each cluster of a clustering result by comparing the objects of a
cluster against all other objects. The limma method is applied to find the
differentially expressed genes.
}
\details{
This function relies on the suggested Bioconductor package 'limma'.
}
\examples{
\dontrun{
data(fingerprintMat)
data(targetMat)
data(geneMat)

MCF7_F = Cluster(fingerprintMat,type="data",distmeasure="tanimoto",normalize=FALSE,
method=NULL,clust="agnes",linkage="flexible",gap=FALSE,maxK=55,StopRange=FALSE)
MCF7_T = Cluster(targetMat,type="data",distmeasure="tanimoto",normalize=FALSE,
method=NULL,clust="agnes",linkage="flexible",gap=FALSE,maxK=55,StopRange=FALSE)

L=list(MCF7_F,MCF7_T)
names=c('FP','TP')

MCF7_FT_DE = DiffGenes(List=L,geneExpr=geneMat,nrclusters=7,method="limma",sign=0.05,
topG=10,fusionsLog=TRUE,weightclust=TRUE,names=names)
}
}
