% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-heatmap.r
\name{distanceheatmaps}
\alias{distanceheatmaps}
\title{Determine the distance in a heatmap}
\usage{
distanceheatmaps(Data1, Data2, names = NULL, nrclusters = 7)
}
\arguments{
\item{Data1}{The resulting clustering of method 1.}

\item{Data2}{The resulting clustering of method 2.}

\item{names}{The names of the objects in the data sets. Default is NULL.}

\item{nrclusters}{The number of clusters to cut the dendrogram in. Default is NULL.}
}
\description{
Internal function of \code{HeatmapPlot}
}
