% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-characterize.r
\name{FeaturesOfCluster}
\alias{FeaturesOfCluster}
\title{List all features present in a selected cluster of objects}
\usage{
FeaturesOfCluster(
  leadCpds,
  data,
  threshold = 1,
  plot = TRUE,
  plottype = "new",
  location = NULL
)
}
\arguments{
\item{leadCpds}{A character vector containing the objects one wants to
investigate in terms of features.}

\item{data}{The data matrix.}

\item{threshold}{The number of objects the features at least should be
shared amongst. Default is set to 1.}

\item{plot}{Logical. Indicates whether or not a BinFeaturesPlot should be
set up. Default is TRUE.}

\item{plottype}{Should be one of "pdf","new" or "sweave". Default is "new".}

\item{location}{If plottype is "pdf", a location should be provided. Default is NULL.}
}
\value{
A list with 2 items: the number of shared features among the objects,
and a character vector of the plotted features.
}
\description{
The function \code{FeaturesOfCluster} lists the number of features objects
of the cluster have in common. A threshold can be set selecting among how
many objects of the cluster the features should be shared. An optional
plot of the features is available.
}
\examples{
\dontrun{
data(fingerprintMat)

Lead=rownames(fingerprintMat)[1:5]

FeaturesOfCluster(leadCpds=Lead,data=fingerprintMat,
threshold=1,plot=TRUE,plottype="new",location=NULL)
}
}
