% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{FindGenes}
\alias{FindGenes}
\title{Find the genes shared across methods}
\usage{
FindGenes(dataLimma, names = NULL)
}
\arguments{
\item{dataLimma}{The output of the \code{DiffGenes} function.}

\item{names}{Optional. Names of the methods. Default is NULL.}
}
\value{
The result is a list with for each cluster a list of the found genes
and the methods that found them.
}
\description{
The \code{FindGenes} function collects the differentially
expressed genes per cluster per method (the output of \code{DiffGenes}) and
determines which genes are found across the methods.
}
\examples{
\dontrun{
MCF7_SharedGenes=FindGenes(dataLimma=MCF7_FT_DE,names=c("FP","TP"))
}
}
