% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{Geneset.intersectSelection}
\alias{Geneset.intersectSelection}
\title{Intersection over resulting gene sets of \code{PathwaysIter} function for a selection of objects}
\usage{
Geneset.intersectSelection(
  list.output,
  sign = 0.05,
  names = NULL,
  seperatetables = FALSE,
  separatepvals = FALSE
)
}
\arguments{
\item{list.output}{The output of the \code{PathwaysIter} function.}

\item{sign}{The significance level to be handled for cutting of the
pathways. Default is 0.05.}

\item{names}{Optional. Names of the methods. Default is NULL.}

\item{seperatetables}{Logical. If TRUE, a separate element is created per
cluster containing the pathways for each iteration. Default is FALSE.}

\item{separatepvals}{Logical. If TRUE, the p-values of the each iteration of
each pathway in the intersection is given. If FALSE, only the mean p-value
is provided. Default is FALSE.}
}
\description{
Internal function of \code{Geneset.intersect}.
}
\examples{
\dontrun{
Geneset.intersectSelection(MCF7_Paths_FandT,sign=0.05)
}
}
