% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.r
\name{HBGF}
\alias{HBGF}
\title{Hybrid Bipartite Graph Formulation}
\usage{
HBGF(
  List,
  type = c("data", "dist", "clust"),
  distmeasure = c("tanimoto", "tanimoto"),
  normalize = c(FALSE, FALSE),
  method = c(NULL, NULL),
  clust = "agnes",
  linkage = c("flexible", "flexible"),
  alpha = 0.625,
  nrclusters = c(7, 7),
  gap = FALSE,
  maxK = 15,
  graphPartitioning = "Spec",
  optimalk = 7
)
}
\arguments{
\item{List}{A list of data matrices. It is assumed the rows are corresponding with the objects.}

\item{type}{Indicates whether the provided matrices in "List" are either data matrices, distance
matrices or clustering results obtained from the data. If type="dist" the calculation of the distance
matrices is skipped and if type="clust" the single source clustering is skipped.
Type should be one of "data", "dist" or "clust".}

\item{distmeasure}{A vector of the distance measures to be used on each data matrix. Should be one of "tanimoto", "euclidean", "jaccard", "hamming". Defaults to c("tanimoto","tanimoto").}

\item{normalize}{Logical. Indicates whether to normalize the distance matrices or not, defaults to c(FALSE, FALSE) for two data sets.}

\item{method}{A method of normalization. Should be one of "Quantile","Fisher-Yates", "standardize","Range" or any of the first letters of these names. Default is c(NULL,NULL) for two data sets.}

\item{clust}{Choice of clustering function (character). Defaults to "agnes".}

\item{linkage}{Choice of inter group dissimilarity (character) for each data set. Defaults to c("flexible", "flexible") for two data sets.}

\item{alpha}{The parameter alpha to be used in the "flexible" linkage of the agnes function. Defaults to 0.625 and is only used if the linkage is set to "flexible".}

\item{nrclusters}{The number of clusters to divide each individual dendrogram in. Default is c(7,7) for two data sets.}

\item{gap}{Logical. Whether the optimal number of clusters should be determined with the gap statistic. Default is FALSE.}

\item{maxK}{The maximal number of clusters to investigate in the gap statistic. Default is 15.}

\item{graphPartitioning}{The graph partitioning method. Currently "Spec" (spectral) is supported.}

\item{optimalk}{An estimate of the final optimal number of clusters. Default is 7.}
}
\value{
The returned value is a list of two elements:
\item{DistM}{A NULL object}
\item{Clust}{The resulting clustering}
The value has class 'Ensemble'.
}
\description{
The \code{HBGF} function implements the Hybrid Bipartite Graph
Formulation, a graph-based consensus clustering method. A bipartite graph is
constructed from the cluster ensemble (objects on one side, clusters on the
other) and partitioned with a spectral method based on the singular value
decomposition followed by k-means.
}
\examples{
\dontrun{
data(mosaic_toy)
L <- mosaic_toy$List
HBGF_toy <- HBGF(List = L, type = "data",
                 distmeasure = c("euclidean", "euclidean"),
                 normalize = c(FALSE, FALSE), method = c(NULL, NULL),
                 clust = "agnes", linkage = c("ward", "ward"),
                 nrclusters = c(7, 7), graphPartitioning = "Spec", optimalk = 7)
}
}
\references{
Fern, X. Z. and Brodley, C. E. (2004). Solving cluster ensemble
problems by bipartite graph partitioning. Proceedings of the Twenty-First
International Conference on Machine Learning (ICML).
}
