% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec.r
\name{HierarchicalEnsembleClustering}
\alias{HierarchicalEnsembleClustering}
\title{Hierarchical ensemble clustering}
\usage{
HierarchicalEnsembleClustering(
  List,
  type = c("data", "dist", "clust"),
  distmeasure = c("tanimoto", "tanimoto"),
  normalize = c(FALSE, FALSE),
  method = c(NULL, NULL),
  clust = "agnes",
  linkage = c("flexible", "flexible"),
  alpha = 0.625
)
}
\arguments{
\item{List}{A list of data matrices. It is assumed the rows are corresponding with the objects.}

\item{type}{indicates whether the provided matrices in "List" are either data matrices, distance
matrices or clustering results obtained from the data. If type="dist" the calculation of the distance
matrices is skipped and if type="clusters" the single source clustering is skipped.
Type should be one of "data", "dist" or"clusters".}

\item{distmeasure}{A vector of the distance measures to be used on each data matrix. Should be one of "tanimoto", "euclidean", "jaccard", "hamming". Defaults to c("tanimoto","tanimoto").}

\item{normalize}{Logical. Indicates whether to normalize the distance matrices or not, default is FALSE. This is recommended if different distance types are used. More details on normalization in \code{Normalization}}

\item{method}{A method of normalization. Should be one of "Quantile","Fisher-Yates", "standardize","Range" or any of the first letters of these names. Default is c(NULL,NULL) for two data sets.}

\item{clust}{Choice of clustering function (character). Defaults to "agnes".}

\item{linkage}{Choice of inter group dissimilarity (character) for each data set. Defaults to c("flexible", "flexible") for two data sets.}

\item{alpha}{The parameter alpha to be used in the "flexible" linkage of the agnes function. Defaults to 0.625 and is only used if the linkage is set to "flexible".}
}
\value{
The returned value is a list of two elements:
\item{DistM}{The resulting distance matrix}
\item{Clust}{The resulting hierarchical structure}
The value has class 'HEC'.
}
\description{
\insertCite{Zheng2014}{MosaiClusteR} proposed the Hierarchical Ensemble Clustering (HEC) algorithm. For each dendrogram, the cophenetic
distances between the object are calculated. The distances are aggregated across the data sets and an ultra-metric which is the closest to the
distance matrix is determined. A final hierarchical clustering is based on the ultra-metric values.
}
\references{
\insertRef{Zheng2014}{MosaiClusteR}
}
