% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-dendro.r
\name{LabelCols}
\alias{LabelCols}
\title{Helper that colours specific dendrogram leaves}
\usage{
LabelCols(x, Sel1, Sel2 = NULL, col1 = NULL, col2 = NULL)
}
\arguments{
\item{x}{A node of a dendrogram (passed by \code{dendrapply}).}

\item{Sel1}{The selection of objects to be colored.}

\item{Sel2}{An optional second selection to be colored. Default is NULL.}

\item{col1}{The color for the first selection. Default is NULL.}

\item{col2}{The color for the optional second selection. Default is NULL.}
}
\value{
The dendrogram node \code{x} with coloured \code{nodePar} and \code{edgePar} attributes.
}
\description{
Internal helper applied via \code{dendrapply} that colours the
leaves and edges of a dendrogram for one or two selections of objects. Used
by \code{LabelPlot}.
}
