% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{PathwayAnalysis}
\alias{PathwayAnalysis}
\title{Pathway analysis with intersection over iterations}
\usage{
PathwayAnalysis(
  List,
  Selection = NULL,
  geneExpr = NULL,
  nrclusters = NULL,
  method = c("limma", "MLP"),
  geneInfo = NULL,
  geneSetSource = "GOBP",
  topP = NULL,
  topG = NULL,
  GENESET = NULL,
  sign = 0.05,
  niter = 10,
  fusionsLog = TRUE,
  weightclust = TRUE,
  names = NULL,
  seperatetables = FALSE,
  separatepvals = FALSE
)
}
\arguments{
\item{List}{A list of clustering outputs or output of the\code{DiffGenes}
function.}

\item{Selection}{If pathway analysis should be conducted for a specific
selection of objects, this selection can be provided here. Default is NULL.}

\item{geneExpr}{The gene expression matrix or ExpressionSet of the objects.}

\item{nrclusters}{Optional. The number of clusters to cut the dendrogram in. Default is NULL.}

\item{method}{The methods for gene and pathway analysis. Default is c("limma","MLP").}

\item{geneInfo}{A data frame with at least the columns ENTREZID and SYMBOL. Default is NULL.}

\item{geneSetSource}{The source for the getGeneSets function, defaults to "GOBP".}

\item{topP}{Overrules sign. The number of pathways to display for each cluster. Default is NULL.}

\item{topG}{Overrules sign. The number of top genes to be returned. Default is NULL.}

\item{GENESET}{Optional. Can provide own candidate gene sets. Default is NULL.}

\item{sign}{The significance level to be handled. Default is 0.05.}

\item{niter}{The number of times to perform pathway analysis. Default is 10.}

\item{fusionsLog}{Logical. To be handed to \code{ReorderToReference}. Default is TRUE}

\item{weightclust}{Logical. To be handed to \code{ReorderToReference}. Default is TRUE.}

\item{names}{Optional. Names of the methods. Default is NULL.}

\item{seperatetables}{Logical. Default is FALSE.}

\item{separatepvals}{Logical. Default is FALSE.}
}
\value{
The intersection over the iterations of the pathway analysis.
}
\description{
The \code{PathwayAnalysis} function performs pathway analysis
multiple times via \code{PathwaysIter} and takes the intersection over the
iterations via \code{Geneset.intersect}.
}
\details{
This function relies on the suggested Bioconductor packages 'MLP', 'biomaRt' and 'org.Hs.eg.db'.
}
\examples{
\dontrun{
MCF7_PathsFandT=PathwayAnalysis(List=L, geneExpr = geneMat, nrclusters = 7,
method = c("limma","MLP"), geneInfo = GeneInfo, geneSetSource = "GOBP",niter=2)
}
}
