% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{PreparePathway}
\alias{PreparePathway}
\title{Prepare data for pathway analysis}
\usage{
PreparePathway(Object, geneExpr, topG, sign)
}
\arguments{
\item{Object}{A clustering output or DiffGenes output element.}

\item{geneExpr}{The gene expression matrix or ExpressionSet of the objects.}

\item{topG}{Overrules sign. The number of top genes. Default is NULL.}

\item{sign}{The significance level to be handled. Default is 0.05.}
}
\value{
A list with the p-values of the genes, the objects and the genes.
}
\description{
The \code{PreparePathway} function prepares the input for the
pathway analysis by computing differentially expressed genes (via limma) for
a selection of objects if these were not provided.
}
\details{
This function may rely on the suggested Bioconductor package 'limma'.
}
\examples{
\dontrun{
PreparePathway(Object=MCF7_F,geneExpr=geneMat,topG=NULL,sign=0.05)
}
}
