% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-navigation.r
\name{SharedComps}
\alias{SharedComps}
\title{Objects shared across clusterings}
\usage{
SharedComps(
  List,
  nrclusters = NULL,
  fusionsLog = FALSE,
  weightclust = FALSE,
  names = NULL
)
}
\arguments{
\item{List}{A list of the clustering outputs to be compared. The first
element of the list will be used as the reference in
\code{ReorderToReference}.}

\item{nrclusters}{If List is the output of several clustering methods, it has
to be provided in how many clusters to cut the dendrograms in. Default is NULL.}

\item{fusionsLog}{Logical. To be handed to \code{ReorderToReference}:
indicator for the fusion of clusters. Default is FALSE.}

\item{weightclust}{Logical. To be handed to \code{ReorderToReference}: to be
used for the outputs of CEC, WeightedClust or WeightedSimClust. If TRUE, only
the result of the Clust element is considered. Default is FALSE.}

\item{names}{Names of the methods or clusters. Default is NULL.}
}
\value{
A list containing the shared objects of all listed elements per cluster.
}
\description{
\code{SharedComps} determines, per cluster, which objects are
shared across all listed clustering results. The clusterings are first
rearranged to a common reference with \code{ReorderToReference} and the
intersection of the lead objects of every method is taken per cluster.
}
\examples{
\dontrun{
data(mosaic_toy)
L <- mosaic_toy$List
res <- list(Cluster(L[[1]], type = "data", distmeasure = "euclidean"),
            Cluster(L[[2]], type = "data", distmeasure = "euclidean"))
Comps <- SharedComps(List = res, nrclusters = 5, fusionsLog = FALSE,
                     weightclust = FALSE, names = c("S1", "S2"))
}
}
