% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{SharedGenesPathsFeat}
\alias{SharedGenesPathsFeat}
\title{Shared genes, pathways and features across methods}
\usage{
SharedGenesPathsFeat(
  DataLimma = NULL,
  DataMLP = NULL,
  DataFeat = NULL,
  names = NULL,
  Selection = FALSE
)
}
\arguments{
\item{DataLimma}{Optional. The output of a \code{DiffGenes} function. Default is NULL.}

\item{DataMLP}{Optional. The output of \code{Geneset.intersect} function. Default is NULL.}

\item{DataFeat}{Optional. The output of \code{CharacteristicFeatures} function. Default is NULL.}

\item{names}{Optional. Names of the methods. Default is NULL.}

\item{Selection}{Logical. Whether a selection of objects was considered. Default is FALSE.}
}
\value{
A list with a Table summarizing the shared elements and a Which
element detailing the shared genes, pathways and features per cluster.
}
\description{
The \code{SharedGenesPathsFeat} function takes the intersection
of the differentially expressed genes, the pathways and the characteristic
features over the methods per cluster.
}
\examples{
\dontrun{
SharedGenesPathsFeat(DataLimma=MCF7_FT_DE,DataMLP=MCF7_Paths_intersection,names=c("FP","TP"))
}
}
