% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{SharedSelection}
\alias{SharedSelection}
\title{Intersection of genes and pathways over multiple methods for a selection of objects.}
\usage{
SharedSelection(
  DataLimma = NULL,
  DataMLP = NULL,
  DataFeat = NULL,
  names = NULL
)
}
\arguments{
\item{DataLimma}{Optional. The output of a \code{DiffGenes} function. Default is NULL.}

\item{DataMLP}{Optional. The output of \code{Geneset.intersect} function. Default is NULL.}

\item{DataFeat}{Optional. The output of \code{CharacteristicFeatures}
function. Default is NULL.}

\item{names}{Optional. Names of the methods or "Selection" if it only
considers a selection of objects. Default is NULL.}
}
\value{
A list with a Table and a Which element for the selection.
}
\description{
Internal function of \code{SharedGenesPathsFeat}.
}
\examples{
\dontrun{
SharedSelection(DataLimma=MCF7_FT_DE,names="Selection")
}
}
