% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{SharedSelectionLimma}
\alias{SharedSelectionLimma}
\title{Intersection of genes over multiple methods for a selection of objects.}
\usage{
SharedSelectionLimma(DataLimma = NULL, names = NULL)
}
\arguments{
\item{DataLimma}{Optional. The output of a \code{DiffGenes} function. Default is NULL.}

\item{names}{Optional. Names of the methods or "Selection" if it only
considers a selection of objects. Default is NULL.}
}
\value{
A list with a Table and a Which element for the shared genes.
}
\description{
Internal function of \code{SharedGenesPathsFeat}.
}
\examples{
\dontrun{
SharedSelectionLimma(DataLimma=MCF7_FT_DE)
}
}
