% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{SharedSelectionMLP}
\alias{SharedSelectionMLP}
\title{Intersection of pathways over multiple methods for a selection of objects.}
\usage{
SharedSelectionMLP(DataMLP = NULL, names = NULL)
}
\arguments{
\item{DataMLP}{Optional. The output of \code{Geneset.intersect} function. Default is NULL.}

\item{names}{Optional. Names of the methods or "Selection" if it only
considers a selection of objects. Default is NULL.}
}
\value{
A list with a Table and a Which element for the shared pathways.
}
\description{
Internal function of \code{SharedGenesPathsFeat}.
}
\examples{
\dontrun{
SharedSelectionMLP(DataMLP=MCF7_Paths_intersection)
}
}
