% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.r
\name{WonM}
\alias{WonM}
\title{Weighting on Membership clustering}
\usage{
WonM(
  List,
  type = c("data", "dist", "clusters"),
  distmeasure = c("tanimoto", "tanimoto"),
  normalize = c(FALSE, FALSE),
  method = c(NULL, NULL),
  nrclusters = list(seq(5, 25), seq(5, 25)),
  clust = "agnes",
  linkage = c("flexible", "flexible"),
  alpha = 0.625
)
}
\arguments{
\item{List}{A list of data matrices. It is assumed the rows are corresponding with the objects.}

\item{type}{Indicates whether the provided matrices in "List" are either data matrices ("data"),
distance matrices ("dist") or clustering results obtained from the data ("clusters").}

\item{distmeasure}{A vector of the distance measures to be used on each data matrix. Should be one of "tanimoto", "euclidean", "jaccard", "hamming". Defaults to c("tanimoto","tanimoto").}

\item{normalize}{Logical. Indicates whether to normalize the distance matrices or not, defaults to c(FALSE, FALSE) for two data sets. More details on normalization in \code{Normalization}.}

\item{method}{A method of normalization. Should be one of "Quantile","Fisher-Yates", "standardize","Range" or any of the first letters of these names. Default is c(NULL,NULL) for two data sets.}

\item{nrclusters}{A list with, for each data matrix, the sequence of numbers of clusters to cut the dendrogram in. Defaults to list(seq(5,25), seq(5,25)).}

\item{clust}{Choice of clustering function (character). Defaults to "agnes".}

\item{linkage}{Choice of inter group dissimilarity (character) for each data set. Defaults to c("flexible", "flexible") for two data sets.}

\item{alpha}{The parameter alpha to be used in the "flexible" linkage of the agnes function. Defaults to 0.625 and is only used if the linkage is set to "flexible".}
}
\value{
The returned value is a list of two elements:
\item{DistM}{The resulting consensus dissimilarity matrix}
\item{Clust}{The resulting clustering}
The value has class 'WonM'.
}
\description{
Weighting on Membership (WonM) is a hierarchy-based consensus
technique. Each data matrix is hierarchically clustered and the resulting
dendrogram is cut into a range of K values. For every cut a co-membership
matrix is built, recording which objects fall in the same cluster. These
matrices are averaged within and across data sets to produce an overall
consensus matrix, which is transformed into a dissimilarity and clustered a
final time.
}
\examples{
\dontrun{
data(mosaic_toy)
fit <- WonM(List = mosaic_toy$List, type = "data",
  distmeasure = c("euclidean", "euclidean"), normalize = c(FALSE, FALSE),
  method = c(NULL, NULL), nrclusters = list(seq(3, 6), seq(3, 6)),
  clust = "agnes", linkage = c("ward", "ward"))
}
}
\references{
Saeed F., Salim N. and Abdo A. (2012). Voting-based consensus clustering for combining multiple clusterings of chemical structures. Journal of Cheminformatics, 4, 37.
}
